refactor: move Go client from client/ to clients/go/ for consistency
Aligns the Go client library location with clients/node/ and clients/python/.
Updates go.work, go.mod module name and replace directive, all import paths,
README references, and CI/release workflows (adds go-client gate job to both).
