+++ .github/workflows/release.yml
@@ -207,6 +207,40 @@ jobs:
         run: python -m mypy queueti/
 
   # ─────────────────────────────────────────────────────────────────────────────
+  # Job 1e: Go client — build and test (gates the release)
+  # Mirrors the go-client job in ci.yml exactly.
+  # ─────────────────────────────────────────────────────────────────────────────
+  go-client:
+    name: "Go client: Build & Test"
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+
+    permissions:
+      contents: read
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - name: Set up Go 1.25.5
+        uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6.4.0
+        with:
+          go-version: "1.25.5"
+          cache-dependency-path: |
+            clients/go/go.sum
+
+      - name: Install Ginkgo CLI
+        run: go install github.com/onsi/ginkgo/v2/ginkgo@v2.28.1
+
+      - name: Build Go client
+        working-directory: clients/go
+        run: go build ./...
+
+      - name: Run Go client tests
+        working-directory: clients/go
+        run: ginkgo ./...
+
+  # ─────────────────────────────────────────────────────────────────────────────
   # Job 2: Build & push Docker image to GHCR
   # Only runs after all four gate jobs pass. Builds a multi-platform manifest
   # (linux/amd64 + linux/arm64) and pushes three conditional tags:
@@ -218,7 +252,7 @@ jobs:
     name: "Docker: Build & push multi-arch image"
     runs-on: ubuntu-latest
     timeout-minutes: 15
-    needs: [backend, frontend, node-client, python-client]
+    needs: [backend, frontend, node-client, python-client, go-client]
 
     permissions:
       contents: read
@@ -354,7 +388,7 @@ jobs:
     name: "PyPI: Publish queue-ti-client"
     runs-on: ubuntu-latest
     timeout-minutes: 10
-    needs: [backend, frontend, node-client, python-client]
+    needs: [backend, frontend, node-client, python-client, go-client]
 
     permissions:
       contents: read
@@ -444,16 +478,16 @@ jobs:
           fetch-depth: 0
 
       - name: Tag client Go sub-module
-        # Creates client/vYYYY.MM.PATCH on the same commit so the client library
+        # Creates clients/go/vYYYY.MM.PATCH on the same commit so the client library
         # is consumable as a Go module:
-        #   go get github.com/Joessst-Dev/queue-ti/client@vYYYY.MM.PATCH
-        # Note: replace directives in client/go.mod are ignored by downstream
-        # consumers, so the local ../  replace does not affect published releases.
+        #   go get github.com/Joessst-Dev/queue-ti/clients/go@vYYYY.MM.PATCH
+        # Note: replace directives in clients/go/go.mod are ignored by downstream
+        # consumers, so the local ../../ replace does not affect published releases.
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git tag "client/${{ github.ref_name }}" "${{ github.sha }}"
-          git push origin "client/${{ github.ref_name }}"
+          git tag "clients/go/${{ github.ref_name }}" "${{ github.sha }}"
+          git push origin "clients/go/${{ github.ref_name }}"
 
       - name: Create GitHub Release
         # --generate-notes  auto-generates release notes from merged PRs since
