+++ clients/go/README.md
@@ -5,13 +5,13 @@ A Go client library for producing and consuming messages from a [queue-ti](https
 ## Installation
 
 ```bash
-go get github.com/Joessst-Dev/queue-ti/client
+go get github.com/Joessst-Dev/queue-ti/clients/go
 ```
 
 Import as:
 
 ```go
-import queueti "github.com/Joessst-Dev/queue-ti/client"
+import queueti "github.com/Joessst-Dev/queue-ti/clients/go"
 ```
 
 ## Quick Start
@@ -416,7 +416,7 @@ import (
     "os"
     "os/signal"
 
-    queueti "github.com/Joessst-Dev/queue-ti/client"
+    queueti "github.com/Joessst-Dev/queue-ti/clients/go"
 )
 
 func main() {
