+++ clients/go/consumer.go
