+++ clients/go/message.go
