+++ clients/go/producer.go
