+++ clients/go/suite_test.go
