+++ clients/go/token_store.go
