+++ go.work
@@ -2,5 +2,5 @@ go 1.25.5
 
 use (
 	./backend
-	./client
+	./clients/go
 )
